+++ java-r5rcore/src/org/ipea/r5r/R5RCore.java
@@ -15,11 +15,8 @@ import org.ipea.r5r.Modifications.R5RFileStorage;
 import org.ipea.r5r.Network.NetworkBuilder;
 import org.ipea.r5r.Process.*;
 import org.ipea.r5r.Utils.Utils;
-import org.rosuda.JRI.RConsoleOutputStream;
-import org.rosuda.JRI.Rengine;
 import org.slf4j.LoggerFactory;
 
-import java.io.PrintStream;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.text.ParseException;
@@ -190,7 +187,7 @@ public class R5RCore {
     }
 
     public void setProgress(boolean progress) {
-        Utils.progress = progress;
+        Utils.setlogProgress(progress);
     }
 
     public void setBenchmark(boolean benchmark) {
@@ -211,6 +208,8 @@ public class R5RCore {
         return Utils.outputCsvFolder;
     }
 
+    public String getLogPath() { return System.getProperty("LOG_PATH"); }
+
     public void setDetailedItinerariesV2(boolean v2) {
         Utils.detailedItinerariesV2 = v2;
     }
@@ -220,10 +219,6 @@ public class R5RCore {
     private static final org.slf4j.Logger LOG = LoggerFactory.getLogger(R5RCore.class);
 
     public R5RCore(String dataFolder, boolean verbose, String nativeElevationFunction) throws Exception {
-        Rengine r = new Rengine();
-        RConsoleOutputStream rs = new RConsoleOutputStream(r, 1);
-        System.setOut(new PrintStream(rs));
-
         if (verbose) {
             verboseMode();
         } else {
@@ -254,7 +249,7 @@ public class R5RCore {
     public RDataFrame detailedItineraries(String fromId, double fromLat, double fromLon, String toId, double toLat, double toLon,
                                                                    String directModes, String transitModes, String accessModes, String egressModes,
                                                                    String date, String departureTime, int maxWalkTime, int maxBikeTime,  int maxCarTime, int maxTripDuration,
-                                                                   boolean dropItineraryGeometry, boolean shortestPath) throws ParseException, ExecutionException, InterruptedException {
+                                                                   boolean dropItineraryGeometry, boolean shortestPath, boolean osm_link_ids) throws ParseException, ExecutionException, InterruptedException {
 
         String[] fromIds = {fromId};
         double[] fromLats = {fromLat};
@@ -266,7 +261,7 @@ public class R5RCore {
         return detailedItineraries(fromIds, fromLats, fromLons, toIds, toLats, toLons,
                 directModes, transitModes, accessModes, egressModes,
                 date, departureTime, maxWalkTime, maxBikeTime, maxCarTime, maxTripDuration,
-                dropItineraryGeometry, shortestPath);
+                dropItineraryGeometry, shortestPath, osm_link_ids);
 
     }
 
@@ -274,7 +269,7 @@ public class R5RCore {
                                                                             String[] toIds, double[] toLats, double[] toLons,
                                                                             String directModes, String transitModes, String accessModes, String egressModes,
                                                                             String date, String departureTime, int maxWalkTime, int maxBikeTime, int maxCarTime, int maxTripDuration,
-                                                                            boolean dropItineraryGeometry, boolean shortestPath) throws ExecutionException, InterruptedException {
+                                                                            boolean dropItineraryGeometry, boolean shortestPath, boolean osmLinkIds) throws ExecutionException, InterruptedException {
         if (Utils.detailedItinerariesV2) {
             // call regular detailed itineraries, based on PointToPointQuery
             FastDetailedItineraryPlanner detailedItineraryPlanner = new FastDetailedItineraryPlanner(this.r5rThreadPool, this.transportNetwork, this.routingProperties);
@@ -285,6 +280,7 @@ public class R5RCore {
             detailedItineraryPlanner.setTripDuration(maxWalkTime, maxBikeTime, maxCarTime, maxTripDuration);
             if (shortestPath) detailedItineraryPlanner.shortestPathOnly();
             if (dropItineraryGeometry) { detailedItineraryPlanner.dropItineraryGeometry(); }
+            if (osmLinkIds) detailedItineraryPlanner.OSMLinkIds();
 
             return detailedItineraryPlanner.run();
         } else {
@@ -680,11 +676,6 @@ public class R5RCore {
 
     public boolean hasFrequencies() {
         return this.transportNetwork.transitLayer.hasFrequencies;
-    }
-
-    public void message(String m) {
-        Rengine r = new Rengine();
-        r.eval("message(\"" + m + "\")");
     }
 
     public void abort() {