+++ r-package/tests/testthat/test-pareto_frontier.R
@@ -195,3 +195,14 @@ test_that("output is correct", {
   expect_true(n_rows <= nrow(origins) * nrow(destinations))
 
 })
+
+
+test_that("using transit outside the gtfs dates throws an error", {
+  expect_error(
+    default_tester(r5r_core,
+           mode='transit',
+           departure_datetime = as.POSIXct("13-05-2025 14:00:00",
+                                           format = "%d-%m-%Y %H:%M:%S")
+    )
+  )
+})