+++ r-package/tests/testthat/test-set_progress.R
@@ -44,16 +44,23 @@ test_that("progress argument works in routing functions", {
 
     progress_regex <- "\\d+ out of \\d+ origins processed\\."
 
-    progress_messages <- capture.output(
-      res <- eval(parse(text = progress_expr)),
-      type = "message"
-    )
+
+    log_file <- file.path(r5r_core$getLogPath())
+    # Clean log before test
+    if (file.exists(log_file)) writeLines("", log_file)
+
+    res <- eval(parse(text = progress_expr))
+    # Wait a bit to ensure Java flushed the logs
+    Sys.sleep(0.2)
+    progress_messages <- readLines(log_file)
     expect_true(any(grepl(progress_regex, progress_messages)))
 
-    non_progress_messages <- capture.output(
-      res <- eval(parse(text = non_progress_expr)),
-      type = "message"
-    )
+    # reset log again
+    writeLines("", log_file)
+    res <- eval(parse(text = non_progress_expr))
+    # Wait a bit to ensure Java flushed the logs
+    Sys.sleep(0.2)
+    non_progress_messages <- readLines(log_file)
     expect_false(any(grepl(progress_regex, non_progress_messages)))
   }